Add pre-release with auto-generated changelog and GHP docs (#51)
This ended up being a bit more than intended; but essentially adds CD
based on Github Actions. Github actions makes it easier to consume
secrets; especially in relation to GitHub. We'll keep Buildkite for now
for regular CI; but it might make sense to move later as well to avoid
having two CI systems.

This doesn't cover regular releases, that'll be a separate ADR so we can
discuss release flows.