Weights&biases logging integration (#104)
Created a new callback WBLogger which is based on the TensorboardLogger
and logs the same things to Weights&Biases.

In order to not make _wandb_ mandatory as a dependency in case someone
doesn't want to use it at all, it is only imported if the flag
`--use_wandb` is set to True. 

Usage notes:
After installing wandb, you can login by running `wandb login` in the
terminal and pasting the API key found in your W&B user settings.

Note that the entity to be used for the wandb initialization (where to
create the runs/projects) is not hardcoded. Instead, it is taken as the
default setting of the user, which should be set to _embark_ for the
work accounts by default. If needed it can be changed either in the W&B
user settings or by passing `entity="embark"` to `wandb.init()`.