Add Storage Subdirectory (#130)
This allows optionally specifying the name of the subdirectory for the
checkpointer and checkpointloader, with the default being "checkpoints".

We also create the "checkpoints" subdirectory from here. 

This is done in emote to avoid code duplication in downstream projects
and match our existing practices, and because 99% of the time, we want
the subfolder to be called "checkpoints".