Agent proxy that supports multiple input keys (#143)
At the moment, we have agent proxies that support only one input key.
These don't support the case of multiple parts in the observation, for
example both features and images.

I wrote a new agent proxy that iterates through the given keys and
creates a dictionary where all parts of the observation are stored with
their respective keys. The policy can then handle these as needed.

Is this a good way to approach this?