Add Conv1d encoder and update Conv2d encoder with option to flatten (#142)
I added a Conv1d encoder which takes as input _(batch size, channels,
length)_, or _(batch size, length, channels)_ if `channels_last=True`,
following the example of the Con2d encoder. I also added the option to
directly add a `nn.Flatten()` layer, instead of having to append one
later on, but the _flatten_ parameter can be set to False if needed.

Also, I updated the existing CURL callback to match the changes done to
the Conv2d encoder.