updating the checkpointer and checkpoint loader (#146)
Made some updates to the Checkpointer and Checkpointloader as follows:

* Removed the extra input arguments for loading networks and optimizers.
Now, you can only load a list of callbacks. However, added an additional
input to the Checkpointloader class. When set to True, this input
instructs the loader to exclusively load the neural network parameters,
disregarding the state_dict of other callbacks. This feature is useful
in scenarios where you need to copy a network from one callback to
another or when you prefer not to load other callback parameters.

Both Checkpointer and Checkpointloader rely on a list of callbacks
rather than a dictionary. To ensure the correct loading of callbacks, we
store a corresponding list of names when saving the checkpoints. These
names serve as references during the loading process. The names are
loaded from the callback names. Callbacks without a name field are
discarded. An error will be raised in case repetitive names are given.