Update windowed scalar (again) (#149)
After fixing the overwriting problem in #148, I found that by logging
using the split key we cause problems in other places where we request
data from the dictionary. By the same logic as #148, when we want to get
logged data we would request the original key, e.g.
"windowed[100]:score" which doesn't exist in the dictionary, so we don't
get any value back. Anything that depends on these values is then
broken.

Instead of splitting the key in a bunch of other places, this code uses
the original key in the dictionary and then splits it only for logging
to the writer.