Simona/add conditions (#162)
- In the Q target updates, terminal masking was used by default, and now
the option is added to deactivate that when needed (+ some updates to
the docstring)
- In the CURL loss calculation, the max was subtracted by default, as in
the original paper. When used together with the temperature variant
(which is different from the original paper), it led to problems with
the gradients and loss not changing. A condition is added now to only
remove the max if not using temperature.