Simona/inits (#163)
The [Pytorch recommendation
](https://discuss.pytorch.org/t/conv-weight-data-vs-conv-weight/83047/2)
is to use weight/bias itself instead of using .data when initializing,
which was used in the past but it was phased out at some point. The
reason they state is that it could cause problems with autograd not
tracking operations correctly, which AFAIK has not caused issues for us,
but it's probably worth changing to keep up with the standard.