Make MemoryProxyWrappers forward methods to wrapee if they do not exist on wrapper (#164)
This PR adds a `MemoryProxyWrapper` base class that all memory proxy
wrappers should inherit from. This base class takes care of forwarding
method calls that do not exist on the wrapper to the inner wrapped
object. This solves problems where one would like to call a method
existing on some inner wrapee but not on the outermost wrapper itself.