Allow temp storage to grow to fit the largest requested data (#174)
Some background for this change: For some AMP code I need to sample the
same memory with two different dataloaders. One loader will always
request less data than the other, and if that is sampled before the
larger one, the larger data no longer fits in the temp storage. This PR
changes so that we grow the temp storage if the data does not fit.
Assuming the requested data sizes do not change over time, this will
only reallocate at the start until the largest buffersize is sampled.

What do you think of this?