Implement Coverage Based Sampling Strategy (#178)
This PR adds a sampling strategy that prioritises samples that haven't
been visited yet over samples that have. The chance of an episode being
sampled is inversely proportional to the amount of times it has been
sampled. This is additionally weighted by an `alpha` parameter, where a
higher alpha means more weighting of less-sampled episodes.

A test is also added to validate that recently added samples have a
higher chance of being chosen.